6. Integrate earlier created components

+++ src/app/auth/auth.service.ts
@@ -1,7 +1,7 @@
-import { Injectable, ɵisObservable } from '@angular/core';
+import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders, HttpErrorResponse} from '@angular/common/http';
 import { SessionStorageService } from './session-storage.service';
-import { BehaviorSubject, catchError, first, from, Observable, throwError } from 'rxjs';
+import { BehaviorSubject, catchError, Observable, throwError } from 'rxjs';
 import { AuthResponse} from './auth.model';
 import { User } from '../user/user.model';
 
@@ -11,45 +11,35 @@ import { User } from '../user/user.model';
 export class AuthService {
 
   private isAuthorized$$: BehaviorSubject<boolean> = new BehaviorSubject<boolean>(false);
-  isAuthorized$: Observable<boolean> = from(this.isAuthorized$$);
+  isAuthorized$: Observable<boolean> = this.isAuthorized$$.asObservable();
 
   constructor(private httpClient: HttpClient, private sessionStorageService: SessionStorageService) { }
 
   login(user: User): void {
-    this.httpClient.post<AuthResponse>('http://localhost:4000', user)
-    .pipe(
-      first(),
-      catchError(this.handleError)
-      )
-      .subscribe(response => {
+    this.httpClient.post<AuthResponse>('http://localhost:4000/login', user)
+      .subscribe(
+        response => {
         this.sessionStorageService.setToken(response.result);
-        this.isAuthorized$$.next(response.successfull);
+        this.isAuthorized$$.next(response.successful);
+      }, 
+      () => {
+        this.isAuthorized$$.next(false);
       })
   }
 
   register(user: User): void {
-    this.httpClient.post('http://localhost:4000', user)
-    .pipe(
-      first(),
-      catchError(this.handleError)
-      );
+    this.httpClient.post<AuthResponse>('http://localhost:4000/register', user).subscribe();
   }
 
   logout(): void {
-    const logoutRequestHeaders = new HttpHeaders().set('Authorization', this.sessionStorageService.getToken());
-    this.httpClient.delete('http://localhost:4000', {headers: logoutRequestHeaders})
-    .pipe(
-      first(),
-      catchError(this.handleError)
-      )
-    .subscribe(() => {
-      this.sessionStorageService.deleteToken()
-    });
-  }
-
-  private handleError(error: HttpErrorResponse) {
-    console.log(error);
-    return throwError('An data error occured, please try again');
+    if (this.sessionStorageService.getToken()) {
+      const logoutRequestHeaders = new HttpHeaders().append('Authorization', this.sessionStorageService.getToken());
+      this.httpClient.delete('http://localhost:4000/logout', {headers: logoutRequestHeaders})
+      .subscribe(() => {
+        this.isAuthorized$$.next(false);
+        this.sessionStorageService.deleteToken();
+      });
+    }
   }
   
 }
